working request limitation. use coherent random samples for subsequent frames

+++ volcanite/data/shader/volcanite/renderer/csgv_bindings.glsl
@@ -159,7 +159,7 @@ layout (std140, binding = 10) uniform render_info {
     uint g_swapchain_index;         // index of this frame in the multiframe swapchain buffer lists
     int g_subsampling;              // border length of the subsampling pixel block in which one sample is rendered
     ivec2 g_subsampling_pixel;      // local coordinate of the currently rendered pixel in the subsampling pixel block
-    float g_random_seed;
+    uint g_random_seed;
 // shading
     float g_factor_ambient;
     float g_light_intensity;
